backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and WebSocket dependencies
openai==1.12.0
websockets==12.0

# Jupyter kernel dependencies
jupyter_client==8.6.0
ipykernel==6.29.0

# Data science dependencies
matplotlib==3.8.2
pandas==2.2.0
numpy==1.26.3
scanpy==1.9.5

PyPDF2==3.0.1
faiss-cpu==1.12.0
tiktoken==0.5.2